+++ CHANGELOG.md
@@ -12,12 +12,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Fold `ControlFlowGraph` and `DataFlowGraph` into `Unit` and `UnitBuilder`.
 - Change `dump()` functions to take a `&Unit` instead of DFG/CFG.
+- Make `FunctionLayout` and `InstLayout` struct private.
+- Make `sig`, `dfg`, `cfg`, and `layout` fields of `UnitData` private.
 
 ### Removed
 - Remove `Function`, `Process`, and `Entity` structs.
 - Remove `Unit` and `UnitBuilder` traits.
 - Remove `Layout` trait.
-- Remove `InstLayout` struct.
 - Remove `func_layout()` and `func_layout_mut()` functions.
 
 ## 0.12.0 - 2020-04-09